Refactored pySearch so it's functionality can be tested. Added initial test and test framework
